requirements.txt
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.7
redis==5.0.0
django-redis==5.3.0
djoser==2.2.0
drf-spectacular==0.26.5
social-auth-app-django==5.3.0
djangorestframework-simplejwt==5.3.0
ruff==0.0.291
drf-nested-routers==0.93.4